[fix] notification action button finished not working in background
Fix 1. action button finished was not working in background, I realized when application is in detached mode, we have to initialize the dependencies like objectbox database and registering get_it singletons again, so I did that and it is working fine. Another change is that instead of passing whole task as notification payload, just pass task id and later get task from objectbox using id.